# openapi.json
{"openapi":"3.1.0","info":{"title":"AI-Powered ATS Resume Screening System","description":"ATS Resume Builder, Analyzer and Job Matching API","version":"1.0.0"},"paths":{"/resume/upload":{"post":{"tags":["Resume"],"summary":"Upload Resume","description":"Upload and extract text from a resume.\nSupports PDF, DOCX and TXT files.","operationId":"upload_resume_resume_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_resume_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Upload Resume Resume Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resume/parse":{"post":{"tags":["Resume"],"summary":"Parse Resume","description":"Extract basic resume sections.","operationId":"parse_resume_resume_parse_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_parse_resume_resume_parse_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resume/health":{"get":{"tags":["Resume"],"summary":"Resume Health","operationId":"resume_health_resume_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/jobs/analyze":{"post":{"tags":["Jobs"],"summary":"Analyze Job","operationId":"analyze_job_jobs_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobDescriptionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/match":{"post":{"tags":["Jobs"],"summary":"Match Resume To Job","operationId":"match_resume_to_job_jobs_match_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobMatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/skills":{"get":{"tags":["Jobs"],"summary":"Available Skills","operationId":"available_skills_jobs_skills_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/ats/score":{"post":{"tags":["ATS Analysis"],"summary":"Ats Score","operationId":"ats_score_api_ats_score_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ATSRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ats/analyze":{"post":{"tags":["ATS Analysis"],"summary":"Ats Analyze","operationId":"ats_analyze_api_ats_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ATSJobRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/builder/preview":{"post":{"tags":["Builder"],"summary":"Preview Resume","operationId":"preview_resume_api_builder_preview_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeBuildRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/builder/improve-bullet":{"post":{"tags":["Builder"],"summary":"Improve Bullet Endpoint","operationId":"improve_bullet_endpoint_api_builder_improve_bullet_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_improve_bullet_endpoint_api_builder_improve_bullet_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/builder/improve-summary":{"post":{"tags":["Builder"],"summary":"Improve Summary Endpoint","operationId":"improve_summary_endpoint_api_builder_improve_summary_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_improve_summary_endpoint_api_builder_improve_summary_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendations/roles":{"post":{"tags":["Recommendations"],"summary":"Recommend Roles","operationId":"recommend_roles_recommendations_roles_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendations/skill-gap":{"post":{"tags":["Recommendations"],"summary":"Skill Gap","operationId":"skill_gap_recommendations_skill_gap_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export/txt":{"post":{"tags":["Resume Export"],"summary":"Export Txt","operationId":"export_txt_export_txt_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export/docx":{"post":{"tags":["Resume Export"],"summary":"Export Docx","operationId":"export_docx_export_docx_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export/pdf":{"post":{"tags":["Resume Export"],"summary":"Export Pdf","operationId":"export_pdf_export_pdf_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ATSJobRequest":{"properties":{"resume_text":{"type":"string","minLength":50,"title":"Resume Text"},"job_description":{"type":"string","minLength":20,"title":"Job Description"}},"type":"object","required":["resume_text","job_description"],"title":"ATSJobRequest"},"ATSRequest":{"properties":{"resume_text":{"type":"string","minLength":50,"title":"Resume Text"}},"type":"object","required":["resume_text"],"title":"ATSRequest"},"Body_improve_bullet_endpoint_api_builder_improve_bullet_post":{"properties":{"bullet":{"type":"string","title":"Bullet"},"target_keywords":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Keywords"}},"type":"object","required":["bullet"],"title":"Body_improve_bullet_endpoint_api_builder_improve_bullet_post"},"Body_improve_summary_endpoint_api_builder_improve_summary_post":{"properties":{"summary":{"type":"string","title":"Summary"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"},"top_skills":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Top Skills"}},"type":"object","required":["summary"],"title":"Body_improve_summary_endpoint_api_builder_improve_summary_post"},"Body_parse_resume_resume_parse_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_parse_resume_resume_parse_post"},"Body_upload_resume_resume_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_resume_resume_upload_post"},"Certification":{"properties":{"name":{"type":"string","title":"Name"},"issuer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Issuer"},"date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date"}},"type":"object","required":["name"],"title":"Certification"},"ContactInfo":{"properties":{"full_name":{"type":"string","title":"Full Name"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"linkedin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linkedin"},"github":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Github"},"portfolio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Portfolio"}},"type":"object","required":["full_name"],"title":"ContactInfo"},"Education":{"properties":{"degree":{"type":"string","title":"Degree"},"institution":{"type":"string","title":"Institution"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"start_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"},"gpa":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gpa"}},"type":"object","required":["degree","institution"],"title":"Education"},"Experience":{"properties":{"title":{"type":"string","title":"Title"},"company":{"type":"string","title":"Company"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"start_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"},"bullets":{"items":{"type":"string"},"type":"array","title":"Bullets"}},"type":"object","required":["title","company"],"title":"Experience"},"ExportRequest":{"properties":{"resume_text":{"type":"string","minLength":50,"title":"Resume Text"},"filename":{"type":"string","title":"Filename","default":"ats_resume"}},"type":"object","required":["resume_text"],"title":"ExportRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobDescriptionRequest":{"properties":{"title":{"type":"string","minLength":2,"title":"Title"},"description":{"type":"string","minLength":20,"title":"Description"}},"type":"object","required":["title","description"],"title":"JobDescriptionRequest"},"JobMatchRequest":{"properties":{"resume_text":{"type":"string","minLength":20,"title":"Resume Text"},"job_description":{"type":"string","minLength":20,"title":"Job Description"}},"type":"object","required":["resume_text","job_description"],"title":"JobMatchRequest"},"Project":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"bullets":{"items":{"type":"string"},"type":"array","title":"Bullets"},"tech_stack":{"items":{"type":"string"},"type":"array","title":"Tech Stack"},"link":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Link"}},"type":"object","required":["name"],"title":"Project"},"RecommendationRequest":{"properties":{"resume_text":{"type":"string","minLength":30,"title":"Resume Text"},"target_role":{"type":"string","title":"Target Role","default":""}},"type":"object","required":["resume_text"],"title":"RecommendationRequest"},"ResumeBuildRequest":{"properties":{"resume":{"$ref":"#/components/schemas/ResumeData"},"template":{"type":"string","title":"Template","default":"classic"}},"type":"object","required":["resume"],"title":"ResumeBuildRequest"},"ResumeData":{"properties":{"contact":{"$ref":"#/components/schemas/ContactInfo"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"skills":{"items":{"type":"string"},"type":"array","title":"Skills"},"education":{"items":{"$ref":"#/components/schemas/Education"},"type":"array","title":"Education"},"experience":{"items":{"$ref":"#/components/schemas/Experience"},"type":"array","title":"Experience"},"projects":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Projects"},"certifications":{"items":{"$ref":"#/components/schemas/Certification"},"type":"array","title":"Certifications"},"achievements":{"items":{"type":"string"},"type":"array","title":"Achievements"},"languages":{"items":{"type":"string"},"type":"array","title":"Languages"},"raw_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Text"}},"type":"object","required":["contact"],"title":"ResumeData","description":"The full, structured representation of a resume — whether it was\ntyped in through the Resume Builder or extracted from an uploaded file."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}